Create README.md
Appendices
Appendix A: Functionality Screenshots
Figure 1.1 - Singing up to the website
 
Upon clicking the sign-up button on the website, the user is taken to the signup form which looks as displayed in the Figure 1.1. , providing users with fields to fill in to create an account.
Figure 1.1.1 – Signing up successfully
 
Upon entering acceptable values into the input forms correctly and signing up the system will display a message indicating that the sign up has succeeded and a verification email has been sent to the entered email address. The account will be created in the database as an unverified account which will be changed once the user clicks the link in the verification email.

1.2 – Signing up without filling in all fields 
 
If the user tries to sign up without filling in all of the fields an error will be displayed on the webpage showing the user that they need to fill in all the fields. Nothing will be entered into the database and the connection will be exited for safety.
1.3 – Invalid email address when attempting to sign up
 
Attempting to signup with an invalid email address will display the error shown in Figure 1.3. the system will throw back the username entered into the username field for convince of the user, however the password fields will be left empty for security reasons.



1.4 – Invalid username when attempting to sign up
 
Attempting to sign up with an invalid username (username with special characters) will result in the system displaying an error message and not saving the account to  the database. The connection will be closed, and the email address will be thrown back to the form for convenience of the user.
1.5 Passwords do not match
 
When attempting to login if the passwords do not match the system will inform the user when trying to register exiting the connection to the database without saving any information and throw back the username and email for user convenience.



1.5.1 – Weak password no uppercase letter
 
If the user tries to register with password without an uppercase the system will not save the account to the database and will inform the user that they must include an uppercase letter this is to improve password strength. The system will keep the username and email in the respective textboxes for ease of use.

1.5.2 – Password Too short when registering
if the password is too short and the user tries to register the system will not allow it and explain to the user that the password must be at least 8 characters. The system will keep the email and username filled in for user convenience. 




1.5.3 – Password must include special character
 
If the user tries to register a password without a special character included the system will not save the account and inform the user that the password must contain a special character. The system will fill in the used email and username for convenience of the user.
1.5.4 – Password must include a number 
 
To improve security of the system will not accept a password if there is no number included. If the user tries to register without a number in the password the system will not save the account and display the error shown in Figure 1.5.4 to the user. The system will fill in the email and username for the user if already filled in for convenience.




1.5.5 – Very weak password indicator
 
1.5.6 - Weak password indicator
 


1.5.7 – Mediocre Password indicator
 


1.5.8 – Strong Password indicator










1.5.9 – Very strong password indicator 
 

1.5.10 – Code snippet for Password meter indicator 
 	
The password meter was created using JavaScript the script calculates a value based on the value in the input field and depending on the strength value the system determines the strength. CSS is then used to display recommendations based on the strength of the password.

1.6.1 – Code snippet for sign up validation
 
A set of if statements are used before an account can be processed and inserted into the database. These check the validity of the input information, if they don’t pass the required parameters then the connection is exited, and the user is sent to the index with a URL error which can be used to display error messages and fill in fields based on the variables available. 
1.6.2 – Code snippet showing how error messages can be displayed
 
Within the html of the actual signup page the php is used to echo out error messages based on the error message thrown to the URL of the webpage as shown in figure 1.6.




1.6.3 – Prepared Statements
 
Prepared statements are used for querying the database to protect the system from being attacked through SQL injections. These essentially allow the system to send and query data separately. This is done by binding parameters and storing results.
2.1 – Encryption used in the system 
 
The system uses the password default encryption method shown in Figure 2.2. This essentially hashes the password using the most efficient algorithm available. The password default function evolves, and the hashing algorithm is rotated based on the most effective method in the current time.




2.2 – Password encryption code snippet
 
The password is hashed by using the password variable passed through the form using the post method into a new variable which is hashed using the algorithm
3.1 – Email verification
 
An email sent to the user with a url to the verification script is sent to a user. This script is protected so only authorised access works. Clicking the link will show take the user to the index page of the site with a message showing that they have been verified
3.2.1 -  Verified Email message 
 
The index page after clicking the email link will have the message displayed in figure 3.2.1




3.2.2 – already verified 
 
If the email has already been verified and the link is accessed the website will inform the user and no action will be taken
3.3 - Logged in unverified
 
Logging in when unverified will show in the home page of the website. This can be used to separate information and access based on whether an account is verified or not.








3.4 – Logged in verified 
 
When logged in and verified the website will display that the user is verified. 

3.5 – Email verification code
 
The following code is used to send an email to the email address of a user who has signed up. Along with the script used to verify a user. The two snippets work together to send a link to a user and to verify a user once the link has been clicked. 
3.6 – Logging in code
 
This code is used to login users to the website the system uses prepared statements to compare the password using the password verify function. It then starts the session and adds session variables which can be called in the website to separate and data based on logged in current user.
3.7 Logging out code
 
Simple code used to log the user out of the system. The session is started, unset and then destroyed and then the user is taken to the index page.